// src/app/books/classes/button.d.ts
export declare class Button {
    iconName: string;
    buttonText: string;
    action?: Function | undefined;
    disabled?: boolean | undefined;
    hidden?: boolean | undefined;
    constructor(iconName: string, buttonText: string, action?: Function | undefined, disabled?: boolean | undefined, hidden?: boolean | undefined)
}